requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
transformers>=4.37
huggingface-hub>=0.20
opencv-python>=4.8
Pillow>=10.1
numpy>=1.26
sentencepiece>=0.1.99
